messages: WIP
Get rid of unnecessary id field from MessageRow. It's equally
efficient to get from ID.
