+++ go/vt/tabletserver/query_executor_test.go
@@ -202,7 +202,6 @@ func TestQueryExecutorPlanInsertMessage(t *testing.T) {
 		Epoch:    0,
 		ID:       sqltypes.MakeString([]byte("1")),
 		Message:  sqltypes.MakeString([]byte("01")),
-		id:       "1",
 	}}
 	if !reflect.DeepEqual(mr, wantmr) {
 		t.Errorf("rows:\n%+v, want\n%+v", got, wantmr)
